refactor(server): isolate totals pandas implem
